requirements.txt
# GCP Cloud Functions dependencies
firebase-admin==6.5.0
google-cloud-pubsub==2.21.1
google-cloud-storage==2.16.0
functions-framework==3.5.0
redis==5.0.4
requests==2.31.0
